Use json-schema for the validation
json-schema is a draf for defining json data structures. The
descriptors now, instead of running validators directly, generate
a json-schema document that is used to validate the value.

The advantage is having a standardized way of params description that
is independent from apipie-params.

We had to drop Type and Proc descriptor, and defining custom
descriptors is pending, but the json-schema supports extensibility.